Moved platform shared code to Platform.h, moved types to Types.h
